fix: remove incompatible py interpreter versions
The implementation of certain harnesses used language features that are
not compatible with legacy Python versions e.g. walrus operator. While
we could ensure that the harnesses are backwards compatible, the legacy
Python versions have already reached their end of life. As such, it
would be better to remove support these versions altogether.